SEMANA06/dia1/requirements.txt
django==3.2
mysqlclient
Pillow
django-ckeditor